tech_event_backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.8.0rc2
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.4.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==6.11.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==2.2.3
